Moving telegram login logic to TelegramLoginService
